% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{The Birthday Function}
\usage{
birthday(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
probability for students to share the same birthday
}
\description{
The Birthday Function
}
\examples{
birthday(20:24)
}
