% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage and Distance Data}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance from a fire}
  \item{DAMAGE}{Damage caused by fire}
}
}
\source{
<FIREDAM.csv from a given file>
}
\usage{
fire
}
\description{
Data of fire damage and distance from a given file
}
\keyword{datasets}
