% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculate Mean and SD}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with mean and sd
}
\description{
This function returns the mean and standard deviation of a numeric vector.
}
