% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Theorem Uniform Solution}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{the number of values per iteration}

\item{iter}{the number of iterations}

\item{a}{the lower limit}

\item{b}{the upper limit}
}
\value{
a histogram of mean values
}
\description{
Central Limit Theorem Uniform Solution
}
\examples{
# mycltu(10,10000,0,10)


}
