% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstfunction.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{My first function}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with components x and y, where y is the square of x
}
\description{
My first function
}
\examples{
myfirstfunction(1:10)
}
