% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Max likelikhood Function}
\usage{
mymaxlikg(lfun, theta)
}
\arguments{
\item{lfun}{A function that computes the log-likelihood for a given parameter value.}

\item{theta}{A numeric vector of candidate parameter values over which to evaluate the likelihood.}
}
\value{
The value of `theta` that maximizes the likelihood function
}
\description{
Computes and plots a one-dimensional likelihood function over a grid of parameter values
and returns the parameter value that maxes the likelihood.
}
\examples{
# Example using a custom log-likelihood function
loglik_fun <- function(theta) sum(dpois(c(2,4), lambda = theta, log = TRUE))
theta_grid <- seq(0.01, 10, length = 100)
mymaxlikg(lfun = loglik_fun, theta = theta_grid)
}
