% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{mynbin function}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{total trials until rth success}

\item{r}{number of successes to be achieved}

\item{p}{probability of success per trial}
}
\value{
the probability P(Y=y)
}
\description{
mynbin function
}
\examples{
# single value:
mynbin(y=10,r=3,p=.4)
# or vector:
mynbin(y=3:12,r=3,p=.4)
}
