% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean value}

\item{sigma}{standard deviation}

\item{a}{upper limit for shaded area}
}
\value{
A list with mu, sigma, a, and the probability P(Y <= a)
}
\description{
myncurve function
}
\examples{
myncurve(mu=1, sigma=2, a=1)
}
