% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Function for optimal ticket sales (both continuous and discrete cases)}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{= number of seats available}

\item{gamma}{= acceptable prob of overbooking}

\item{p}{= prob a ticket holder shows up}
}
\value{
a list containing
\itemize{
   \item \code{nd} - optimal tickets, discrete)
   \item \code{nc} - optimal tickets, continuous)
   \item \code{N} - number of seats
   \item \code{p} - probability of showing up
   \item \code{gamma} - probability of overbooking
 }
}
\description{
Function for optimal ticket sales (both continuous and discrete cases)
}
\examples{
result <- ntickets(N = 400, gamma = 0.02, p = 0.95)
result
}
